lots of work on stuff

+++ include/gwat/mcmc_sampler_internals.h
@@ -42,7 +42,7 @@ public:
 	double **step_prob;
 	double **prob_boundaries;
 	double *chain_temps;
-	double **swap_partners; 
+	int **swap_partners; 
 	bool random_swaps = true;
 	//########
 	double **chain_neighborhoods;
